# [#120379] Improve performance of all transaction view