--- a/app/models/account.rb
+++ b/app/models/account.rb
@@ -12,6 +12,7 @@ class Account < ActiveRecord::Base
 
   has_many   :account_users, :inverse_of => :account
   has_one    :owner, :class_name => 'AccountUser', :conditions => {:user_role => AccountUser::ACCOUNT_OWNER, :deleted_at => nil}
+  has_one    :owner_user, :through => :owner, :source => :user
   has_many   :business_admins, :class_name => 'AccountUser', :conditions => {:user_role => AccountUser::ACCOUNT_ADMINISTRATOR, :deleted_at => nil}
   has_many   :price_group_members
   has_many   :order_details
@@ -120,9 +121,9 @@ class Account < ActiveRecord::Base
     account_number <=> obj.account_number
   end
 
-  def owner_user
-    self.owner.user if owner
-  end
+  # def owner_user
+  #   self.owner.user if owner
+  # end
 
   def owner_user_name
     owner_user.try(:name) || ""