--- a/spec/factories/nufs_accounts.rb
+++ b/spec/factories/nufs_accounts.rb
@@ -3,7 +3,7 @@ overridable_factory :nufs_account do
     "9#{'%02d' % n}-7777777" # fund3-dept7
   end
 
-  sequence(:description, 'a') { |n| "nufs account #{n}" }
+  sequence(:description, 'aaaaaaaa') { |n| "nufs account #{n}" }
   expires_at { Time.zone.now + 1.month }
   created_by 0
 end