The loading of the search options, especially when running cross-facility
has gotten fairly slow. This makes some changes to how some of those
options are loaded to speed things up.

While testing against UIC, we discovered they're running MySQL 5.1, so some
of the optimizations aren't as good as they are on 5.6+.

##### Cross Facility view:

MySQL version | Original | Optimized
----|----|----
5.1 | ~2.75s | 1.4s
5.6 | ~1.5s  | 1s

##### Single facility:

MySQL version | Original | Optimized
----|----|----
5.1 | ~1.5s | 1.4s
5.6 | .9s | .7s

And from testing on staging (which is 5.1), I saw similar scale improvements:

Facility view | Original | Optimized
----|----|----
Cross | 2s | .75ms
Single | .8s | .75ms

I also tested against Oracle for single facilities, and saw slight, but
similar improvements.