@jhanggi 2015-09-22
I've cherry-picked this into a UIC branch that I released to their staging environment for testing. I'd still like to hear any feedback or at least get a thumbs up.