@drench 2015-09-23
Just a few minor comments. There are style update possibilities here and there too, but as far as query changes go it looks all right to me.