@jhanggi 2015-09-23
Updated to remove the extra `to_a`