@jhanggi 2015-09-24
There was an odd NufsAccount ordering issue that appears to stem from alphabetization of descriptions, so I changed it from going a..z...aa...zz to aaaaaaa...aaaaaaaz...aaaaaaza...zzzzzzz. The tests are now passing.